Allow search label to use different sizes
This adds the ability to pass in a size - `xl`, `l`, `m`, and `s` - to
change the size of the label's text.

To avoid a breaking change this doesn't use the label component. Using
the label component would mean replacing the search label's current
default styles with the `govuk-label` defaults.

Adds tests and documentation.

Linting updates to label component.