Load Youtube video instantly when cookies accepted
YoutubeLinkEnhancement checks if the user has accepted all cookies, and
progressively enhances a youtube link into an embedded video if that
condition is met.

This check occurs when the module is initialised on page load.

Therefore, if a user who has not yet accepted cookies lands on a page
containing a YouTube video, then clicks "Accept cookies" in the cookie
banner in order to watch said video, they would have to refresh the page
in order for the video to load.

This is not ideal and can lead to confusion or frustration.

I have introduced an additional check for the case in which cookies have
not been accepted on initialisation, which listens for the
'cookie-consent' event and re-initialises the YoutubeLinkEnhancement
functionality if the event was detected.
As a result, the video should load as soon as the user has accepted
cookies, and not on the next page load.