Add account layout wrapper component
Some account-related components, contents, and CSS are repeated across
different apps on GOVUK.

The first instance of this was when we relocated the account homepage
from the govuk-account-manager-app to the frontend app, which meant that
some of the page furniture and content was duplicated between those apps.

More recently, the accounts team are working towards the goal of single
page notifications which will enable users who have linked their email
notifications address and their GOVUK account email address to manage
their email subscriptions through their account. Part of this work is
making the email management pages look the same as the account when the
user is authenticated with a linked account (these pages being rendered
through the email-alert-frontend app).
That would make it three separate apps where the same page furniture and
content are repeated.

The purpose of this PR is to nip this issue in the bud by creating an
account layout/wrapper component which can be shared between different
apps. This will mean that the content, the CSS, the markup for all the
different building blocks that make up the account layout will live in
the gem which should help us prevent inconsistencies between different
applications using this layout.