Extend layout for public with account components
Update layout_for_public component to support GOVUK account components.
Existing variations of the component will remain unchanged, but a few
options have been introduced to support the main GOVUK account layout
variations.

The intention is for this to be used with a layout template in static,
to apply account-like page furniture to pages that should have the
account layout, while not being rendered by the
govuk-account-manager-prototype app.