Fix `line-height` on step-by-step nav header
Applying a custom `line-height` on a span element doesn't have any effect, as span is an inline element. We can set the property on the parent h2 element  instead, as it's a block element. This fixes an issue on mobile where the browser default h2 `line-height` is inherited by the span.