Fix layout error that occurs when resizing browser window
When moving from mobile to desktop view, and resizing the window to
exactly 769 pixels wide the layout was broken. This was because the
listener to change the layouts was added at this breakpoint, but the
function wouldn't be fired until the _next_ resize event one pixel
later.

Since this was a `resize` listener being fired by another resize or
matchMedia listener it could be removed and replaced with firing the
function.