Fix global banner wrapper size
The `global-banner-wrapper` element is rendered whether or not there is an active global banner as it's a slot for other elements such as the survey banner. When there is no content in this container, because of the negative margin-top it pulls the element below (often the back link or the breadcrumbs) higher up than it should be. By setting a `min-height` we make sure no layout alteration happen in this scenario, while it scales up when content is presented within the container.