Fix document list children
Class composition (in this case `gem-c-document-list__children`  with `govuk-list`) means that when we have the same attribute (let's say `margin-top`) the order of precedence is dictated by where these class definitions are presented in the compiled CSS files – known as document order. Document order makes styling flakey. When swapping a component in and out or changing the order of SCSS imports we may end up with different styling. This happened in https://github.com/alphagov/finder-frontend/pull/2616 when switching to the new `gem_layout`. To overcome this issue we drop class composition in favour of redefining the only two attributes that are not overridden by `gem-c-document-list__children`.