Reorder the navigation lists vertically
Lists are easier to scan when each column is alphabetically sorted vertically.
This is possible to do automagically with CSS columns, but Chrome doesn't like
CSS columns when links are using `text-decoration-thickness` [1][2] - so another
solution was needed.

CSS Grid Layout has support going back to IE 10[3] and allows for vertical
columns as long as the number of items in the columns are known - so when the
navigation is updated the CSS controlling the grid needs to be updated.

[1]: https://github.com/alphagov/govuk-frontend/issues/2204
[2]: https://bugs.chromium.org/p/chromium/issues/detail?id=1190987
[3]: https://caniuse.com/css-grid