Add mixin for more robust navigation updates
The mixin outputs the styles needed for the grid layout to work in IE10+ - this
is quite a repetative syntax, so the mixin takes the number of items and the
number of columns and works out the rest.

Added comments to the pertinent parts of the navigation localisation to remind
future people that they need to update the styles as well.