Add two word connector to locale files
The `to_sentence` method uses different values to join sentences with
two items compared to sentences with three or more items.

We were previously defining the three or more, but not the value for
when there are two items. This resulted in falling back to the default
word for the language. Where the language did not have a translation for
the other words (e.g. "Applies to"), we ended up with a non-English word
in the middle of an English sentence (e.g. "Applies to England a Wales"
instead of "Applies to England and Wales" on the Welsh translation, as
there is currently no Welsh version of the other words).