Accept cookie consent from Digital Identity query parameter
We're agreed with Digital Identity (DI) that we'll pass cookie consent
between our two domains with query parameters on links and redirects.
This is so that users don't see multiple banners when they move
between the different account management pages, as some will be on
GOV.UK and others will be on the DI domain.  The Privacy team have
approved sharing of consents.

The parameter is called `cookie_consent`, and takes one of three
values:

- `accept`: the user has consented to usage cookies on the other
domain, so we should persist the consent on this domain.

- `reject`: the user has rejected usage cookies on the other domain,
so we should persist the consent on this domain.

- `not-engaged`: the user has not engaged with the cookie banner (or
some other consent mechanism) on the other domain, so we should not
change things on this domain.

In practice we can ignore any value other than `accept` or `reject`,
as those are the only two which trigger a change.

Co-Authored-By: Dana Cotoran <Dana.Cotoran@digital.cabinet-office.gov.uk>