Add cookie_consent parameter to links to Single Sign On
We want to share a cookie consent between www.gov.uk and account.gov.uk
(or whatever the DI domain ends up being).  Rather than try to tie this
to clicking the accept / reject button in the banner, we've decided to
go for the simpler approach of passing a query parameter around.

If a user hasn't engaged with the cookie banner and clicks on a link to
the other domain, add `cookie_consent=not-engaged`.

If a user has rejected usage cookies and clicks on a link to the other
domain, add `cookie_consent=reject`.

Otherwise, if a user has accepted usage cookies and clicks on a link to
the other domain, add `cookie_consent=accept` and the GA client ID if
available.