Update input with search icon CSS
There is an issue with the search icon variant of the input component,
where Chrome user agent for "autofill selected" (when a user selects one
of the autofill options suggested by Chrome) makes the search icon
disappear.

This is happening because the search icon is applied using the
background property in CSS, which is being overruled by CSS from the
user agent stylesheet.

After fruitless attempts to override the browser's autofill styling, I
landed on the current solution which is applying the search icon as a
separate element before the input field, rather than as a background
image directly on the input.

Bonus: this also fixes a bug wherein the icon and input field are
misaligned when magnification/zoom text only is used in Firefox