Remove coronavirus priority taxons
Now that the coronavirus hub pages have been unpublished, we no longer
want content tagged to a coronavirus child taxon to have the "Part Of"
breadcrumb.

The priority taxon code is quite tangled with the Brexit priority
taxons. For expediency, a light touch has been taken to remove the
coronavirus specific code. Some thought needs to be given to whether the
concept of priority taxons is still need at all outside of the Step by
Step space.