Remove coronavirus priority taxon tests
Now that the coronavirus hub pages have been unpublished, we no longer
want content tagged to a coronavirus child taxon to have the "Part Of"
breadcrumb.

Now that the coronavirus taxons are no longer considered a "priority" taxon we
need to remove the tests that check for the taxons to appear in contextual navigation.