Update navigation header layout on resize
Previously, the layout was updated when shifting from mobile <--> desktop screen
sizes. This assumed that the menu only changed size when the media query
breakpoint was reached - however, the contents of the drop down menu affected
the height before this breakpoint was reached. This updated the layout to be
checked and updated on every resize rather than just when the media query
breakpoint is reached.