Fix layout jank when on slow connection
The styles for the navigation header used the global `js-enabled` class to tweak
the layout for when JavaScript was available. The `js-enabled` class was added
to the `<body>` element by an inline script, which meant that the class was
added as soon as the inline script was parsed.

If the components JavaScript wasn't yet loaded and/or initialised, this would
cause the component layout to be broken until the JavaScript was loaded, parsed,
and run.

Instead of using the global `js-enabled` class the navigation header component
now adds it's own `js-module-initialised` class to the component to signify that
the layout changes needed for the JavaScript layout to work can be used. This
class is added _after_ the component has been successfully initialised, so
prevents any layout jank caused by the race between the component JavaScript and
the global `js-enabled` class.