Update application root for RTM interaction
We use `Rails.root` quite a lot in Rails Translation Manager to figure
out the calling applications root directory. However, for
govuk_publishing_components, this was being reported as
`govuk_publishing_component/spec/dummy` due to the Gem not being a true
Rails app and implementing it's own 'dummy' app which was overwriting
the root dir.

As `Rails.root` needs to set as `.../spec/dummy` to keep Govuk
Publishing Components Rake tasks and tests operational, we've monkey
patched `root` which switches Rails root directory to the real root of
the Gem: `.../govuk_publishing_components`, if the caller is
`rails_translation_manager`.

This is a bit of a hack but does mitigate some confusing root behaviour
within the Gem, and means we don't have to pass in paths to the various
rake tasks in RTM, which would've been a chore and prone to errors.