Amend consent cookie expiration date
When the cookie banner 'reject' option is selected, the
cookies_preferences_set cookie is created and set to expire after 365
days. However, the expiration date is not adjusted on the pre-existing
cookie_policy cookie (which is always set to a default value, unless
already present).

This commit updates the cookie banner `rejectCookieConsent` function, so
that when cookies are explicitly rejected, the consent cookie is set to
its default value, and its expiration date is reset.