Amend if condition around cookie_consent parameter
The explicit cross domain links script has some logic to decorate
cross-domain links with a cookie_consent=reject parameter if usage
cookies have been rejected.
To determine if usage cookies were rejected, it checks for the
`cookies_policy` cookie: `if (cookieConsent.usage === false){...}`

A `cookies_policy` cookie should always exist, because a check is
performed on page load (via cookie banner initialisation script) so
that, if it is not already present, `cookies_policy` is created with
the default values.

However, as a preventative measure against a scenario where the
`cookies_policy` cookie is not present in  the browser, it is a good
idea to update the condition from `if (cookieConsent.usage === false)`
to `if (cookieConsent && cookieConsent.usage === false)` which will
not throw an `Uncaught TypeError` if the cookie is not found.