Fix font-size on super navigation header
Rem unit measures are not meant to be used in compatibility mode (in conjunction with legacy dependencies such as govuk_template) as the font-size for the root element is not controlled. Hence we wrap these declarations and only apply them when rem units are set to be used by `govuk-frontend`.