Add page margin back when screen size allows
The navigation header needs to be smaller when below 360px wide - so the chevron
is removed and the right side page margin is removed. Once above 360px wide
these can both be added back in - but only the chevron was, which resulted in an
ever so slightly lopsided header. This adds the margin back.