Adjust navigation header black bar height
This changes the height of the navigation header to match the GOV.UK Frontend
header.

In the far distant past and for reasons unknown the black bar in the header
component was adjusted to be a different height to the height of the GOV.UK
Frontend / Design System header component. This meant that when a user goes from
www.gov.uk to a service, the header would change height - which is not the
seamless experience that we're aiming for.