Add organisation colour for DLUHC
DLUHC = Department for Levelling Up, Housing and Communities

The colour has a contrast ratio of 14.75, so not adding a websafe colour
for this organisation.

The same change has been made in [govuk-frontend](https://github.com/alphagov/govuk-frontend/pull/2434),
however we need to also make the change here until the next release of
govuk-frontend.