Remove use of `govuk-list` from navigation header
The only thing that was not being override was `list-style`, so this declaration
was added to the class in the component and `govuk-list` was removed.

The use of `govuk-list` was slightly problematic - the class was being appearing
after the component styles in some applications due to the application CSS being
later in the cascade that the CSS from Static - so the component spacing would
be overridden.