Amend explicit-cross-domain-links js code
Amend explicit-cross-domain-links code to account for the possibility of
someone accepting/rejecting cookies on the page that the cross-domain
link is present. On detecting the cookie consent or rejection event, the
cross-domain link module should re-run and append the correct parameters
to any cross domain links present on the page.