Apply track click attribute to wrapper
Apply the data-module="gem-track-click" for the single page notification
button to a component wrapper div.
The reason behind applying this to a wrapper and not to the form is that
the form gets swapped out with the form returned from the
personalisation endpoint, so the track click module would have to be
re-started on the new, updated form.
Applying the track-click attribute to a wrapper instead means we don't
have to re-start the gem-track-click module for the updated form.