Update Start button with info text
The info text not associated with the Start button when present.  If you provide information text ("info_text") to a button, the text is rendered underneath the button but not linked to it in any way. 

This means it won't be read out to screenreader users when they focus on the button, and they could miss important information.

Creating an ID on both elements should the info text be present and pairs them together[1] then also hides the info text from assistive tech as this has already been read out.

Using example content to demo:
 "Start now Sometimes you want to explain where a user is going to"

[1]https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/ARIA_Techniques/Using_the_aria-labelledby_attribute