Add margin_bottom option to success alert
The GOV.UK sucess alert component has a fixed bottom margin which is
fine in most use cases of this component.

Recently, however, the success alert component has been added to
the top of publication pages whose content has a large margin on the
top. This makes it so that the visual space between the success alert
and the content of the page is excessive, hence the need to introduce
this option to the component.