Adjust explicit-cross-domain-links.js
It has transpired that the explicit-cross-domain-links.js script was not
set up correctly which is breaking our cross domain tracking currently.

The source of the issue is the ga parameter which contains a timestamp
which actually expires and becomes invalid after two minutes.

The way the script worked previously was that the relevant cross-domain
links and forms on the page were decorated with the _ga and
cookie_consent parameters on page load. Doing it on page load is an
issue due to the time-sensitive nature of the _ga parameter, as
mentioned above.

This changes the way the script works so that instead of decorating
crossdomain links and forms with additional parameters on page load, we
decorate them when they are interacted with.
Besides ensuring that the _ga param is always valid and up-to-date, this
also makes it so that there is no longer a need to listen for cookie
banner events, so I also removed the cookie-reject event as part of this
commit.