Fix flash of unpersonalised content
Fix flash of unpersonalised content that can happen to the Single page
notification button when someone is logged in and has subscribed to
notifications on the current page.

This feature has been designed with the personalisation bit as a JS
enhancement. A JS request is made to a personalisation endpoint on page
load, which returns a response based on which the button reflects the
subscription status of the current user to the current page.

The swap from old button to updated button should happen in a near
instant in ideal conditions, but can be more obvious in certain
instances (such as if someone's internet connection is not good)

This aims to work around this problem by initially hiding the button
when JS is enabled, then making it visible immediately when XHR has
completed.