Add exemption list for COVID CTA by document type
A new COVID CTA (call to action) is being created to be added to the
contextual_sidebar to be rendered on most pages to help users find out
how to book their booster vaccination.

There are a few exemptions. The CTA won't be displayed on pages of a
particular document_type.

The CTA component will be added in a later commit.