Add exemption list for COVID CTA by taxon
A new COVID CTA (call to action) is being created to be added to the
contextual_sidebar to be rendered on most pages to help users find out
how to book their booster vaccination.

There are a few exemptions. The CTA won't be displayed on pages tagged
to particular taxons.

The CTA component will be added in a later commit.