Metadata tweak for "See all updates" interaction
On some publications, the metadata component "See all updates" link is
an in-page link which jumps to the list of updates at the bottom of the
page. The reason for this is that the list of notes can be quite long
and so it would not be ideal to have this list right at the top of the
publication page as even when hidden behind a visibility toggle, such a
list could take a lot of scrolling to get past and inconvenience some
users.

There was some feedback in UR that suggested that people found it
confusing that an additional click is required to show the list of
updates (i.e after clicking "See all updates" at the top, one must click
"see all updates" again at the bottom; the expected behaviour is that
the list of update notes should be made visible straight away after the
click at the top).

This is a workaround in order to achieve the intended outcome, which
adds a simple script for the metadata component which fires a click
event on the toggle trigger when the "See all updates" link is clicked.