Add check to big number to convert plus suffixes to subscript elements
If a value passed to a big number is a string and ends with a plus symbol, strip it out and re-amend it to the end of the original value as a span.

This allows us to safely amend the positioning of the plus symbol so that it sits at the baseline of the big number value, adding to govuk's visual harmony.