Upgrade visual regression test action
This upgrades the visual regression test action to take advantage of the new
(non-deprecated) Percy CLI and Percy-Capybara gem upgrade. It only required a
few changes, but if you're trying to run this locally it will need the removal
of the `node_modules` directory as there seemed to be confusion between the old
and new when running `npx percy ...`.