Port cards component from `frontend`
The team created a local cards component for the homepage in  `frontend`. It is currently used in the list of topics and government activity sections on the homepage.

We want to reuse it for Mainstream Browse pages so it needs to be made available to other apps via the gem.

Port the code over from  `frontend` so that it can be modified for inclusion in the gem.