Configure jasmine-browser
jasmine-browser-runner uses an underlying Express server that is placed in in the srcDir (in our case public/assets). We need to prevent these NodeJS files from being added to the page we test with, which is why we use this exclusion syntax.

For reference, below is the error we receive if the folder is not excluded

Suite error: undefined
  Message:
    Uncaught ReferenceError: require is not defined
  Stack:
    ReferenceError: require is not defined
        at http://localhost:35455/__src__/express/lib/application-b5249205605770867132eb65e251e899e30502bed0f5db6a3e3cefe09ec533bd.js:17:20