Use GitHub Actions for Continuous Integration
This commit configures GitHub Actions to run continuous integration tasks. It also removes the Jenkins configuration file, which was what previously handled CI.

It uses a matrix to run tests against both versions of Ruby that this gem supports: 2.6 and 2.7. It also handles publishing new versions of the gem to RubyGems.org.