Calculate rows with CSS grid. Remove IE11 polyfill
The frontend repo version of the component uses the `columns` mixin which polyfills IE11. However, this requires that the number of items in the grid is set in scss. This approach worked for the limited use cases in frontend, but wouldn't work for the new topics templates which will have an unknown number of items in the grid.

Set the grid styles directly on the component instead of using the `columns` mixin - we briefly discussed amending the mixin to cope with this new use case but it would have required quite extensive changes to the mixin, and wouldn't have provided a lot of value to the cards component which would have only really needed a couple of grid properties set by the mixin and modifying or omitting large chunks of the rest of the mixin.

Change the component to calculate the number of rows using `grid-template-rows: auto;`, instead of needing to hardcode this in the scss.

Use `grid-auto-rows` to use the height of the tallest cell in the grid to set row height.

For browsers that don't support CSS grid (we're mainly concerned with IE11), fall back to a single column but constrain it to 50% width on desktop - this was recommened by our designer to improve usability especially for users of screen magnifiers. Set the width to 50% for all browsers, and then reset it back for browsers that support CSS grid.