Use ruby installed chromedriver
The webdrivers gem installs chromedriver in the ~/.webdrivers directory
as part of bundle install. If we add this directory to the $PATH then we
can avoid the need to have a separate installation of chromedriver.

It strikes me that there could potentially be some scenarios where this
is flakey. I know that webdrivers can be configured with an env var
(WD_INSTALL_DIR) so could be installed somewhere else, however I think
if it's set up special the only bad thing will be that you get told you
need to install chromedriver.