Split jasmine task into jasmine:ci and jasmine:browser
It seemed convenient to have a quick ability to run jasmine tests in the
browser. It also seemed like it'd be confusing if the old command
"jasmine" ran tests in the console and "jasmine:browser" ran them in the
browser. So I've switched "jasmine" to be "jasmine:ci" for consistency.