Update webdrivers before running jasmine
As we expect to use a webdrivers installation of chromedriver it makes
sense to ensure it is installed before we try make use of it.

The webdrivers gem seems to install chromedriver rather magically, I'm
not too sure if it's installed automatically as part of bundle running
(part of just the gem being included) or part of specific tasks. I think
it's simpler to run this manually rather than relying on assumptions.