Don't use webdrivers for yarn chromedriver
In the time since PR 2625 [1] was merged we've learnt that the webdrivers
gem is going to cause a problem for the GOV.UK development environment,
govuk-docker, when running on M1 macs.

We have decided to change tact and focus on installing a chromedriver on
the underlying machine rather than using webdrivers as a source of this
software [2] [3]. This seems extra pragmatic now that we have Javascript
dependencies also using chromedriver.

A step towards this is removing this intertwining of Ruby webdriver
dependency from this repo's yarn tests.

The effect merging this will have is:

- The build will continue to pass in GitHub Actions CI - chromedriver is
  installed
- Users running jasmine tests in govuk-docker will need to be using [2]
- Users running jasmine tests on their local machine will have to
  install chromedriver manually (there is a helpful prompt)

[1]: https://github.com/alphagov/govuk_publishing_components/pull/2625
[2]: https://github.com/alphagov/govuk-docker/pull/577
[3]: https://github.com/alphagov/govuk_test/pull/43