Add support for Rails 7
This commit removes the "< 7" version constraint on the Rails dependency, making it compatible with Rails 7 applications.

It also updates the Gemfile.lock to run on Rails 7. This means that the Component Guide that are served as part of this gem will now run on Rails 7 – including the [standalone version][1] hosted on Heroku.

[1]: https://components.publishing.service.gov.uk