Use Psych 4 for YAML parsing
[Psych][1] is the default YAML parsing library in Ruby. It's included as part of Ruby's standard library, but it's also [available as a gem][2].

Different versions of Ruby include different versions of Psych:

| Psych | Ruby           |
| ----- | -------------- |
| 4.0.3 | 3.1.0 .. 3.1.1 |
| 3.3.2 | 3.0.3          |
| 3.3.0 | 3.0.0 .. 3.0.2 |
| 3.1.0 | 2.6.0 .. 2.7.5 |

The full table is available here: https://stdgems.org/psych/

The jump from Psych 3 (in Ruby 2.7 and 3.0) to Psych 4 (in Ruby 3.1) introduced breaking changes in the way YAML is loaded. In Psych 3 `YAML.load_file` would load YAML files without any safety features enabled – whereas in Psych 4 `YAML.load_file` is an alias of `YAML.safe_load_file`. To achieve the same behaviour in Psych 4, you need to use the new `YAML.unsafe_load_file` method.

The YAML files we load in this gem are trusted, and they use special features such as aliases which are deemed 'unsafe', so we need to use Psych 4's `YAML.unsafe_load_file` method.

Whilst it's possible to load YAML differently depending on the Psych version available (see alphagov/govuk-connect#69), this approach isn't ideal due to the additional noise it adds to code.

Instead, this commit simply adds Psych 4 as a dependency of the project. This dependency will override the older version included in the Ruby standard library, meaning we end up with Psych 4 regardless of the Ruby version we're running on.

[1]: https://github.com/ruby/psych
[2]: https://rubygems.org/gems/psych