Install yarn dependencies before gem build
This is to correct an issue where we built a version of the gem (28.7.0)
but it lacked the NPM modules and was not a usable release (we have
since yanked it).

I checked whether this had an impact by comparing file size of the gem
before and after this was applied.

Before:

-rw-r--r--  1 runner docker 723K Mar  1 18:56 govuk_publishing_components-28.7.0.gem

After:

-rw-r--r--   1 runner docker 2.2M Mar  1 19:04 govuk_publishing_components-28.7.0.gem

As previous releases of this gem were ~ 2.2MB this now looks to have
corrected the issue.