Add support for both Psych 3 and Psych 4
The method to load YAML files is different between Psych 3 and Psych 4.

Ruby 2.7 and 3.0 include Psych 3 in their standard library, whereas Ruby 3.1 includes Psych 4.

Since this gem needs to be compatible with Ruby >= 2.7, we need to be able to cope with having either Psych 3 or Psych 4 available in our environment.

Whilst it would have been possible to require Psych 4 in this gem's dependencies, it would have meant that every downstream application woudl also be bumped to Psych 4 even if they were expecting Psych 3.