For reference: remove tests that don't test template logic
Remove some component template tests that were added in the same PR.

There is currently an undocumented convention with GOV.UK Publishing components to only test markup when the template contains explicit logic (eg. conditionals or forks) for generating different types of markup, rather than testing the integrity of the generated markup eg. that the correct classes and elements are generated (unless there's logic involved).

We do have visual regression tests and do PR reviews so those give us additional confidence.

However we might wish to consider testing for the integrity of the markup in the future, especially if we saw instances where this could have flagged unwanted changes.

Leave these removed tests out as a separate commit in case we want to revisit this later.