Update testing
Update testing to include regex. 

Capybara 3 now fails tests with text with whitespace. It's more strict as the hidden text renders as white space so fails.

There is a suggested temporary update  to include "Capybara.default_normalize_ws = true" as a more global update however this might suggested this disappears in 4.