Don't use embedded Ruby for rum-loader
Mixing Ruby and JavaScript can cause a variety of challenges and limits
the portability of the script. This script has been changed to instead
use a data attribute on the script tag rather than needing embedded
Ruby.

The immediate problem I had to resolve with this is that this didn't
work properly with Sprockets 4. However as mixing Ruby and JavaScript is
an undesirable approach I felt it best to decouple this.