Provide compatibility with Sprockets 4
This makes a few minor modifications that allows an application to make
use of Sprockets 4.

This configuration will allow govuk_publishing_components to be used
both in Sprockets 3 and 4 applications (the only blocker I found for
Sprockets 4 was that component_guide/application.css wasn't available,
which was resolved by adding it the asset configuration).