Try this one trick to remove seg faults
Using Sassc and Sprockets seems somewhat prone to segmentation faults
which seems to be a common issue with Sprockets 4. As per
https://github.com/rails/sprockets/issues/633 and
https://github.com/sass/sassc-ruby/issues/207, switching concurrent
asset generation seems to resolve the problem. The unproven theory is
that this reflects that Sprockets 4 uses threads which causes problems
for sassc.

A consequence of this is that asset pre-compilation may be slower than
it could be.