Catch errors when rendering the chart
MagnaCharta requires tables to use table header elements in order to
render the chart correctly.

The error will now be caught if the govspeak markdown used is not valid.

If an error does occur, both the chart and the toggleLink button will
not be rendered and the table will be shown by default.

Set ENABLED flag to false if stacked chart table missing headers

As the table headers are required in order to render the chart correctly,
setting the ENABLED flag to false will prevent the chart from being
contructed and also the toggle button from being displayed.