Update cross domain linking script to use init
It was noted that cross-domain linking between the account and the DI
domain stopped working around 25th March.
Around that time, a lot of good work was done around removing jQuery
from GOVUK. As part of that work there were some changes to the script
which starts the modules in govuk_publishing_components where support
for legacy modules was removed.

The explicit-cross-domain-links script was still using start() to
initialise. This changes it to init().