Make navbar links have same styling if JS disabled
This commit fixes an issue where the styling for the links in the navbar
that are visible when JS is not enabled where not resembling the
dropdown toggle buttons that appear in the navbar when JS is enabled.

The pseudo underline that appears on hover has been updated to allow the
width to be set. This fixes the issue where the pseudo underline would
be too short for links. To fix focus styling issues, the text of the
link is now wrapped in a span. This allows the background colour of the
link to be set without padding and means that the pipe height can be
set independently of the link height.