Add HTTP protocol measurement to RUM
Knowing which HTTP protocol is being used is important for understanding how the
different protocols affect load times and our user's experiences.

This adds in a new measurement to the payload that will be sent to Speedcurve's
LUX service - this will allow us to compare the different HTTP protocols and
how they affect our performance metrics.