Fix HTTP protocol reporting in Safari
Monitoring the type of HTTP connection that a user's using will allow us to see
how the different HTTP protocols affect performance.

This used `DOMContentLoaded` to make sure that `performance.getEntriesByType()`
was not undefined. However, this was not reporting anything in Safari. It
appears that the event listener was being added after the event was fired.

Checking the `readyState` of the document lets us see whether the document is
loading or ready - and then, depending on the state, either the `load` event has
a listener attached or the measurement is done immediately.