Moved emergency_banner component to the gem
As part of the plan to retire static, the emergency banner
component is being moved to govuk_publishing_components. This
will reduce applications reliance on static as emergency banner
is used on every application that uses the public layout. The
result of this is that frontend applications will no longer have
to include static and govuk_publishing_components to import globally
used components.

This commit includes the docs, html and scss that will be removed from static
once the emergency banner has been implemented in the gem. The classes
have had the 'gem-c-' prefix added in the stylesheets and html. The print
style for emergency banner has also been added in the same directories
as other print styles in the gem.