Add gem-track-click to specific sections of layout_footer
Previously gem-track-click was added to the parent element of layout_footer. As we need to add
seperate gem-track-clicks to the copyright links that are generated from the locale files in order to
track them, this results in double events when the copyright links were clicked. Also added
data-track-links-only because when non links were being clicked it was triggering a tracked click event.